NNFX Trading Strategy Template
This contains the NNFX trading strategy templates for tradingview version 5. The blank template is to be used to populate in the tutorial on youtube. The full template shall be used to follow along and validate your work when watching the tutorial. Once users are comfortable, they can use the full template to insert their own indicators for their own strategy.